+++ src/zandronum/src/g_strife/a_acolyte.cpp
@@ -33,7 +33,7 @@ DEFINE_ACTION_FUNCTION(AActor, A_HideDecepticon)
 		return;
 	}
 
-	EV_DoDoor (DDoor::doorClose, NULL, self, 999, 8*FRACUNIT, 0, 0, 0);
+	EV_DoDoor (DDoor::doorClose, NULL, self, 999,(int)( 8*FRACUNIT), 0, 0, 0);
 	if (self->target != NULL && self->target->player != NULL)
 	{
 		P_NoiseAlert (self->target, self);
